requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
pillow==11.3.0
psycopg2==2.9.10
PyJWT==2.10.1
pymongo==3.11.4
python-dotenv==1.1.1
pytz==2025.2
sqlparse==0.5.3
daphne
